Optimize for OOP Design
