+++ src/Customer.java
@@ -1,5 +1,6 @@
 import java.io.*;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
@@ -38,13 +39,15 @@ class Customer extends Account {
         this.membership = membership;
     }
 
+
+
     public static void registerAccount(String filename) throws IOException, NoSuchAlgorithmException {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter a username:");
         String username = scanner.nextLine();
         Scanner sc = new Scanner(System.in);
         do {
-            if (Method.ifUsernameExisted("E:\\Study\\order-management-system\\Data\\account.txt", username)) {
+            if (ifUsernameExisted("E:\\Study\\order-management-system\\Data\\account.txt", username)) {
                 username = sc.nextLine();
             } else {
                 break;
@@ -72,8 +75,105 @@ class Customer extends Account {
         address = Method.validateEmpty(address);
 
         Customer account = new Customer(Method.generateID("C", filename), username, password, fullName, phoneNumber, email, address);
-        Method.writeToDatabase(account, filename);
+        writeToDatabase(account, filename);
         System.out.println("Register successfully!");
+    }
+
+    public static boolean ifUsernameExisted(String fileName, String username) {
+        try {
+            BufferedReader br = new BufferedReader(new FileReader(fileName));
+            String line;
+            while((line = br.readLine()) != null) {
+                String[] data = line.split(","); // split line by comma delimiter
+
+                if (username.equals(data[1])) {
+                    System.out.println("Username existed! Please choose another one!");
+                    return true;
+                }
+            }
+            br.close(); // close the BufferedReader object
+
+        } catch (FileNotFoundException e) {
+            throw new RuntimeException(e);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        return false;
+    }
+
+    public static Customer updateCustomerInformation(Customer customer) throws IOException {
+        Scanner sc = new Scanner(System.in);
+        System.out.printf("You are changing your personal information for %s!\n", customer.getUsername());
+        System.out.printf("Your old name is '%s', please input your new name or type your old!\n", customer.getFullName());
+        String newName = sc.nextLine();
+        customer.setFullName(newName);
+        System.out.printf("Your old phone number is %s, please input your new one or type your old!\n", customer.getPhoneNumber());
+        String newPhoneNumber = sc.nextLine();
+        customer.setPhoneNumber(newPhoneNumber);
+        System.out.printf("Your old address is %s, please input your new one or type your old!\n", customer.getAddress());
+        String newAddress = sc.nextLine();
+        customer.setAddress(newAddress);
+        System.out.printf("Your old email is %s, please input your new one or type your old!\n", customer.getEmail());
+        String newEmail = sc.nextLine();
+        customer.setEmail(newEmail);
+
+        Path path = Paths.get("E:\\Study\\order-management-system\\Data\\account.txt");
+        List<String> lines = Files.readAllLines(path);
+
+        // Replace the customer information if it reaches the id
+        for (int i = 0; i < lines.size(); i++) {
+            String[] fields = lines.get(i).split(",");
+            if (fields[0].equals(customer.getId())) {
+                fields[3] = customer.getUsername();
+                fields[4] = customer.getPhoneNumber();
+                fields[5] = customer.getEmail();
+                fields[6] = customer.getAddress();
+                lines.set(i, fields[0] + "," + fields[1] + "," + fields[2] + "," + fields[3] + "," + fields[4] + "," + fields[5] + "," + fields[6] + "," + fields[7]);
+                break;
+            }
+        }
+
+        // Write the modified lines back to the text file
+        Files.write(path, lines);
+        return customer;
+
+    }
+
+    public static Customer updatePassword(Customer customer) throws IOException {
+        Scanner sc = new Scanner(System.in);
+        String username = customer.getUsername();
+        System.out.printf("You are changing your passowrd for %s!\n", customer.getUsername());
+        System.out.printf("Please input your old password\n", customer.getFullName());
+        String oldPassword = sc.nextLine();
+
+        Path path = Paths.get("E:\\Study\\order-management-system\\Data\\account.txt");
+        List<String> lines = Files.readAllLines(path);
+
+        // Replace the customer information if it reaches the id
+        for (int i = 0; i < lines.size(); i++) {
+            String[] fields = lines.get(i).split(",");
+            if (fields[1].equals(username) && fields[2].equals(oldPassword)) {
+                do {
+                    System.out.println("Your password is correct, please input the new password");
+                    String newPassword = sc.nextLine();
+                    System.out.println("Please re-input your new password");
+                    String newRePassword = sc.nextLine();
+                    if (newPassword.equals(newRePassword)) {
+                        fields[2] = newPassword;
+                        lines.set(i, fields[0] + "," + fields[1] + "," + fields[2] + "," + fields[3] + "," + fields[4] + "," + fields[5] + "," + fields[6] + "," + fields[7]);
+                        break;
+                    }
+
+                } while (true);
+
+                break;
+            }
+        }
+        // Write the modified lines back to the text file
+        Files.write(path, lines);
+        System.out.println("Password changed successfully!");
+        return customer;
+
     }
 
     @Override
@@ -137,6 +237,18 @@ class Customer extends Account {
 
     public void setMembershipPlatinum() {
         this.membership = "Platinum";
+    }
+
+    public static void writeToDatabase(Customer account, String filename) {
+        try {
+            FileWriter fw = new FileWriter(filename, true);
+            BufferedWriter bw = new BufferedWriter(fw);
+            bw.append(account.getId()).append(",").append(account.getUsername()).append(",").append(account.getPassword()).append(",").append(account.getFullName()).append(",").append(account.getPhoneNumber()).append(",").append(account.getEmail()).append(",").append(account.getAddress()).append(",").append(account.getMembership()).append("\n");
+            bw.close(); // close the BufferedWriter object
+            fw.close();
+        } catch (IOException e) {
+            System.out.println("Error writing to database file: " + e.getMessage());
+        }
     }
 
     public static List<Customer> listCustomers() throws IOException, NoSuchAlgorithmException {
